Update PlusMinus.java
Given an array of integers, calculate the ratios of its elements that are positive, negative, and zero. Print the decimal value of each fraction on a new line with  places after the decimal. In example:
arr=[1,1,0,-1,-1]  There are n=5 elements, two positive, two negative and one zero. Their ratios are 2/5=0.400000, 2/5=0.400000, 1/5=0.200000
 Results are printed as:

0.400000
0.400000
0.200000